// src/forum/posts/interfaces/post-response.interface.ts
import { IAuditFields } from "src/common/interfaces/common.interface";

export interface IPostResponse extends IAuditFields{
    id: string;
    title: string;
    slug: string;
    content: string;
    userId: string;
    author: string;
    authorNickname: string;
    authorBio: string;
    authorRole: string;
    
    topicId: string;
    topicTitle: string;
    topicSlug: string;
    topicTagId: string;

    parentId?: string;
    parentTitle?: string;
    parentSlug?: string;
    parentContent?: string;
    postCount?: number;
    viewCount: number;

    lastPostAt?: Date;

    parentUserId?: string;
    parentAuthor?: string;
    parentAuthorNickname?: string;
    parentAuthorBio?: string;
    parentAuthorRole?: string;
    status: boolean;
}